# docs/mkdocs/api/swagger.yaml
basePath: /
definitions:
...
  api.committeeJSON:
    properties:
      created_at:
        type: string
      creation_signature:
        type: string
      id:
        type: string
      lead_address:
        type: string
      members:
        items:
          $ref: '#/definitions/api.committeeMemberJSON'
        type: array
      threshold_m:
        type: integer
      total_n:
        type: integer
    type: object
  api.committeeMemberJSON:
    properties:
      address:
        type: string
      moniker:
        type: string
      pub_key_b64:
        type: string
    type: object
  api.committeeRequestJSON:
    properties:
      creation_signature:
        type: string
      lead_address:
        type: string
      members:
        items:
          $ref: '#/definitions/api.committeeMemberJSON'
        type: array
      threshold_m:
        type: integer
      total_n:
        type: integer
    type: object
...
  api.errorBody:
    properties:
      code:
        type: string
      invariants:
        items:
          $ref: '#/definitions/api.invariantResultJSON'
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
  api.errorEnvelope:
    properties:
      error:
        $ref: '#/definitions/api.errorBody'
    type: object
...
  api.invariantResultJSON:
    properties:
      invariant:
        type: string
      ok:
        type: boolean
      reason:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Chain launch coordination server. Coordinators and validators authenticate
    via secp256k1 challenge-response (ADR-036)
  title: ChainCoord API
  version: "1.0"
paths:
...
  /committee/{launch_id}:
    get:
      parameters:
      - description: Launch UUID
        in: path
        name: launch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.committeeJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: Get committee
      tags:
      - committee
...
  /launch/{id}/committee:
    post:
      consumes:
      - application/json
      description: Replaces the committee on a DRAFT launch. Lead coordinator only.
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Committee definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.committeeRequestJSON'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.committeeJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Set committee
      tags:
      - committee
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Session token obtained from POST /auth/verify. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"